Set offset for terminator searching to use the terminator bytesize
If a multi-byte line terminator is used (e.g. "\r\n") and a read attempt
into the buffer gets only part of the multi-byte terminator, then this
logic would get two lines together.

More importantly, if this terminator was at the very end of a POP3
message (i.e. end of message last line is ".\r\n") then another read
would be attempted for data which would never come. This would result in
a Net::ReadTimeout.